revised but unfinished
