+++ understand_sentiment/dataprovider.py
@@ -16,9 +16,10 @@ from paddle.trainer.PyDataProvider2 import *
 
 def hook(settings, dictionary, **kwargs):
     settings.word_dict = dictionary
-    settings.input_types = [
-        integer_value_sequence(len(settings.word_dict)), integer_value(2)
-    ]
+    settings.input_types = {
+        'word': integer_value_sequence(len(settings.word_dict)),
+        'label': integer_value(2)
+    }
     settings.logger.info('dict len : %d' % (len(settings.word_dict)))
 
 
@@ -32,4 +33,4 @@ def process(settings, file_name):
             word_slot = [
                 settings.word_dict[w] for w in words if w in settings.word_dict
             ]
-            yield word_slot, label
+            yield {'word': word_slot, 'label': label}
